pr ready, commented some code that is not supposed to be releaaed yet, adjusted some tests and added some source code in the targets
